requirements.txt
httpx
pyyaml
pandas
python-dateutil